requirements.txt
django>=4.2
djangorestframework
django-cors-headers
django-markdownify
requests
python-dotenv
django-bootstrap4
django-debug-toolbar
